createAttributes
